+++ .eslintrc
@@ -15,6 +15,7 @@
     "no-bitwise": ["off"],
     "no-console": ["off"],
     "no-extra-parens": ["off"],
+    "no-unused-vars": ["error", { "vars": "all", "args": "after-used", "ignoreRestSiblings": false }],
     "one-var": ["error", {
       "var": "always",
       "let": "always",
@@ -26,7 +27,9 @@
     "space-infix-ops": ["error", { "int32Hint": false }],
     "space-unary-ops": ["off"]
   },
+  "parser": "babel-eslint",
   "parserOptions": {
-    "ecmaVersion": 8
+    "ecmaVersion": 8,
+    "sourceType": "module"
   }
 }
